cmd/multiping: work around rate-limit (?)
On an remote APU board, pinging 35 hosts, I've noticed missing outgoing
packets. I assume that either the next-hop (an Aruba switch), the onboard
NIC, or the Linux kernel itself rate-limits the number of outging ICMP
packets.

This needs further investigation. As naïve workaround, we delay executing
the goroutine by calling `time.Sleep` for a few ms and hope for the best.